added basic implementation for interactive hsv panel, TODO remove force reload button and implement auto-refresh for the image on hsv-change

+++ src/state/app_state.rs
@@ -40,7 +40,7 @@ pub struct GalleryState {
 #[derive(Clone, Copy)]
 pub struct HSVState {
     pub panel_visible: Signal<bool>,
-    pub hue: Signal<f64>,
-    pub saturation: Signal<f64>,
-    pub value: Signal<f64>,
+    pub hue: Signal<f32>,
+    pub saturation: Signal<f32>,
+    pub value: Signal<f32>,
 }
